tests reafctored
