+++ blocks/shared-styles/_children/promos/medium/mock-data.js
